requirements.txt
alembic==1.11.1
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Deprecated==1.2.14
executing==1.2.0
Flask==2.3.2
Flask-Cache==0.13.1
Flask-Caching==2.0.2
Flask-captcha2==2.0.1
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
Flask-Migrate==4.0.4
Flask-Session==0.5.0
flask-shell-ipython==0.5.1
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
limits==3.5.0
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyMySQL==1.0.3
python-dotenv==1.0.0
redis==4.5.5
requests==2.31.0
rich==13.4.2
six==1.16.0
SQLAlchemy==2.0.16
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
WTForms==3.0.1
